pyproject.toml
[project]
name = "ai-risk-control-agent-for-mt5"
version = "0.2.0"
description = "Risk-rule + LLM evaluator over MT5 broker snapshots."
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.100.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "tortoise-orm>=0.21.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
app = "app.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
